+++ src/app/entities/room.ts
