+++ src/app/shared/components/dark-button/dark-button.component.ts
@@ -1,7 +1,7 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { NgIf } from '@angular/common';
-import { IconDatabase } from '../../../helpers/icons-database';
+import { IconDatabase } from '../../helpers/icons-database';
 
 @Component({
   selector: 'app-dark-button',
