+++ src/app/shared/components/message/message.component.html
@@ -1,21 +1,21 @@
 <div [ngClass]="{
-  'justify-end':messageType == 'user' && isCurrentUser(),
+  'justify-end':messageType == 'user' && isCurrentUser,
   'justify-center':messageType == 'system'
   }" class="flex flex-row mb-2.5 ">
   <div>
     <section [ngClass]="{
-    'bg-message1/20':messageType == 'user' && isCurrentUser(),
-    'bg-message0/20':messageType == 'user' && !isCurrentUser(),
-    'border-message1':messageType == 'user' && isCurrentUser(),
-    'border-message0':messageType == 'user' && !isCurrentUser(),
-    'rounded-br-lg':messageType == 'user' && isCurrentUser(),
-    'rounded-bl-lg':messageType == 'user' && !isCurrentUser(),
-    'ml-auto':messageType == 'user' && isCurrentUser(),
+    'bg-message1/20':messageType == 'user' && isCurrentUser,
+    'bg-message0/20':messageType == 'user' && !isCurrentUser,
+    'border-message1':messageType == 'user' && isCurrentUser,
+    'border-message0':messageType == 'user' && !isCurrentUser,
+    'rounded-br-lg':messageType == 'user' && isCurrentUser,
+    'rounded-bl-lg':messageType == 'user' && !isCurrentUser,
+    'ml-auto':messageType == 'user' && isCurrentUser,
     'max-w-[33dvw] w-fit rounded-20 py-2 px-5 gap-0.5 border-2': messageType !='system',
     }" class="mb-0.5">
       <p class="text-16-20 font-normal">{{ content }}</p>
     </section>
-    <section *ngIf="messageType != 'system'" [ngClass]="{'ml-auto':messageType == 'user' && isCurrentUser()}"
+    <section *ngIf="messageType != 'system'" [ngClass]="{'ml-auto':messageType == 'user' && isCurrentUser}"
              class="flex flex-row gap-2.5 items-baseline">
       <p class="text-14-17 font-normal text-muted">{{ sentAt }}</p>
       <p *ngIf="messageType == 'user'" class="text-15-18 font-semibold text-muted">{{ senderUsername }}</p>
