+++ src/app/shared/helpers/icons-database.ts
