+++ src/app/widgets/room/room.component.html
